Fixed Lwmte's merge. Scancode does not works on my notebook, keysym works correct. Must be updated shift and ctrl const in config. Fixed character controller functions.

+++ controls.h
@@ -100,8 +100,8 @@ typedef struct control_settings_s
 
     int8_t   joy_axis_map[AXIS_LASTINDEX];      // Axis array for action mapper.
 
-    int16_t  action_map[ACT_LASTINDEX];         // Actions array for action mapper.
-    int16_t  action_alt[ACT_LASTINDEX];         // Alternate actions array for alternate controls.
+    int32_t  action_map[ACT_LASTINDEX];         // Actions array for action mapper.
+    int32_t  action_alt[ACT_LASTINDEX];         // Alternate actions array for alternate controls.
 
 }control_settings_t, *control_settings_p;
